Registra Compras json ejemplo: {     "nroCompra": 4,     "cantidadExistente": 15,     "bodegaBean": {         "id": 1,         "descripcionBodega": ""     },     "articuloBeanCompras": [         {             "id": 1,             "codigoArticulo": null,             "nombreArticulo": null,             "precio": null         },         {             "id": 2,             "codigoArticulo": null,             "nombreArticulo": null,             "precio": null         }     ] }

+++ backEnd/src/main/java/com/bo/openlogics/sales/model/DetalleComprasArticulo.java
@@ -18,7 +18,9 @@ import java.math.BigDecimal;
 public class DetalleComprasArticulo implements Serializable {
 
     @Column(name="CANTIDAD_EXISTENTE")
-    private BigDecimal cantidadExistente;
+    private Integer cantidadExistente;
+
+
 
     private DetalleComprasArticuloPK pk = new DetalleComprasArticuloPK();
 
@@ -27,11 +29,11 @@ public class DetalleComprasArticulo implements Serializable {
     public DetalleComprasArticulo() {
     }
 
-    public BigDecimal getCantidadExistente() {
+    public Integer getCantidadExistente() {
         return cantidadExistente;
     }
 
-    public void setCantidadExistente(BigDecimal cantidadExistente) {
+    public void setCantidadExistente(Integer cantidadExistente) {
         this.cantidadExistente = cantidadExistente;
     }
